Restructure Packages. Add interfaces for stack and queue menus
