+++ src/Lab3/Menus/QueueMenus/LinkedQueueMenu.java
@@ -1,23 +1,23 @@
-package Lab3.Menus;
+package Lab3.Menus.QueueMenus;
 
 import Lab3.behavior.AppLoop;
 
-import Lab3.implementations.Stack.LinkedStack;
+import Lab3.implementations.Queue.LinkedQueue;
 
-import Lab3.interfaces.StackMenuInterface;
+import Lab3.interfaces.QueueMenuInterface;
 
 import java.util.Scanner;
 
-public class LinkedStackMenu implements StackMenuInterface {
+public class LinkedQueueMenu implements QueueMenuInterface {
     private Scanner scanner;
     private AppLoop appLoop;
     private boolean flag = true;
-    private LinkedStack stack;
+    private LinkedQueue linkedQueue;
 
-    public LinkedStackMenu(Scanner scanner, AppLoop appLoop) {
-        this.appLoop = appLoop;
+    public LinkedQueueMenu(Scanner scanner, AppLoop appLoop) {
         this.scanner = scanner;
-        this.stack = new LinkedStack(1);
+        this.appLoop = appLoop;
+        this.linkedQueue = new LinkedQueue();
     }
 
     @Override
@@ -26,17 +26,17 @@ public class LinkedStackMenu implements StackMenuInterface {
         String[] commandsList = input.split(" ");
 
         switch (commandsList[0]) {
-            case "push" -> {
+            case "enqueue", "en" -> {
                 this.flag = false;
                 if (commandsList.length > 1) {
                     for (int i = 1; i < commandsList.length; i++) {
-                        this.stack.push(commandsList[i]);
+                        this.linkedQueue.enqueue(commandsList[i]);
                     }
                 }
                 else {
                     while (true) {
                         String input1 = takeElementInput(this.scanner);
-                        this.stack.push(input1);
+                        this.linkedQueue.enqueue(input1);
                         System.out.println("MORE ELEMENTS? Y/N");
                         input1 = this.scanner.nextLine();
                         if (input1.equalsIgnoreCase("n")) {
@@ -45,61 +45,55 @@ public class LinkedStackMenu implements StackMenuInterface {
                     }
                 }
             }
-            case "pop" -> {
+            case "dequeue", "d" -> {
                 this.flag = false;
-                Object peekedElement = this.stack.peek();
-                if (peekedElement != null) {
-                    System.out.println("POPPED ELEMENT: " + this.stack.peek());
-                    this.stack.pop();
+                Object rearElement = this.linkedQueue.getRearElement();
+                if (rearElement != null) {
+                    System.out.println("POPPED ELEMENT: " + this.linkedQueue.getRearElement());
+                    this.linkedQueue.deque();
                 }
 
             }
             case "peek" -> {
                 this.flag = false;
-                Object peekedElement = this.stack.peek();
+                Object peekedElement = this.linkedQueue.peek();
                 if (peekedElement != null) {
-                    System.out.println("LAST ELEMENT IN THE STACK: " + peekedElement);
+                    System.out.println("LAST ELEMENT IN THE QUEUE: " + peekedElement);
                 }
             }
             case "full", "f" -> {
                 this.flag = false;
-                System.out.println("STACK:\n" + this.stack.toString());
+                System.out.println("QUEUE:\n" + this.linkedQueue.toString());
             }
             case "status" -> {
                 this.flag = false;
-                int sizeStack = this.stack.getLinkedListStack().length;
-                this.stack.setOccupiedSpace();
-                int occupiedSpace = this.stack.getOccupiedSpace();
-                System.out.print("STACK IS: ");
-                if (this.stack.isEmpty()) {
-                    System.out.println("EMPTY: 0 / " + sizeStack);
-                }
-                else if (occupiedSpace == sizeStack){
-                    System.out.println("FULL: " + occupiedSpace + " / " + sizeStack + " ELEMENTS");
+                System.out.print("QUEUE IS: ");
+                if (this.linkedQueue.isEmpty()) {
+                    System.out.println("EMPTY");
                 }
                 else {
-                    System.out.println("PARTIALLY FULL: " + occupiedSpace + " / " + sizeStack + " ELEMENTS");
+                    System.out.println("FULL");
                 }
             }
             case "search", "s" -> {
                 this.flag = false;
                 if (commandsList.length > 1) {
                     for (int i = 1; i < commandsList.length; i++) {
-                        this.stack.search(commandsList[i]);
+                        this.linkedQueue.search(commandsList[i]);
                     }
                 }
                 else {
                     String searchedElement = takeElementInput(this.scanner);
-                    this.stack.search(searchedElement);
+                    this.linkedQueue.search(searchedElement);
                 }
             }
             case "empty" -> {
                 this.flag = false;
-                if (!this.stack.isEmpty()) {
-                    this.stack.empty();
+                if (!this.linkedQueue.isEmpty()) {
+                    this.linkedQueue.empty();
                 }
                 else {
-                    System.out.println("NO ELEMENTS IN THE STACK!");
+                    System.out.println("NO ELEMENTS IN THE QUEUE!");
                 }
             }
             case "help", "h" -> {
@@ -107,8 +101,9 @@ public class LinkedStackMenu implements StackMenuInterface {
                 printHelp();
             }
             case "exit", "e" -> {
+                this.linkedQueue.deleteQueue();
                 printQuit();
-                this.appLoop.setActiveMenu(new StackMenu(this.scanner, this.appLoop));
+                this.appLoop.setActiveMenu(new QueueMenu(this.scanner, this.appLoop));
             }
             default -> {
                 printInvalid();
@@ -127,6 +122,6 @@ public class LinkedStackMenu implements StackMenuInterface {
 
     @Override
     public void printGreetings() {
-        System.out.println("WELCOME TO LINKED LIST STACK IMPLEMENTATION MENU!");
+        System.out.println("WELCOME TO LINKED LIST QUEUE IMPLEMENTATION MENU!");
     }
 }
